// apps/trade-pro/src/pages/defineItemPos/type.ts
export type TItemHistoryTable = {
  RecordNo: number;
  Id: number;
  ItemCode: string;
  ItemName: string;
  BaseUnitId: number;
  ItemCategoryId: number;
  ItemTypeId: number;
  ItemClassId: number;
  PurchaseGLAC: number;
  SaleGLAC: number;
  COGSGLAC: number;
  HSCode: string;
  PackQty: number;
  ItemStatus: boolean;
  ManufactureId: number;
  ManufacturePartNo: string;
  BuyerSupplierId: number;
  BuyerPartNo: string;
  ProductNo: null | number;
  MinStockLevel: number;
  OptimalStockLevel: number;
  MaxStockLevel: number;
  ReorderLevel: number;
  ItemOriginId: number;
  CostPrice: number;
  UOMScheduleIdCostRate: number;
  PurchasePrice: number;
  UOMScheduleIdPurRate: number;
  RetailPrice: number;
  UOMScheduleIdRetailRate: number;
  WholeSalePrice: number;
  UOMScheduleIdWhsRate: number;
  AmountCalcType: null | string | number;
  BarcodeImage: string | ImageBitmap;
  BarcodeNo: number;
  BaseCropYear: string | number;
  ApplyGST: boolean;
  ApplyVAT: boolean;
  ApplyExciese: boolean;
  OtherTax: boolean;
  Pic1: null | ImageBitmap;
  Pic2: string | ImageBitmap;
  PostDate: Date | number;
  PostUser: number;
  PostState: boolean;
  OrganizationId: number;
  CompanyId: number;
  BranchesId: number;
  ProjectsId: number;
  GlStockAccountTitle: string;
  GlSaleAccountTitle: string;
  GlCgsAccountTitle: string;
  CategoryDescription: string;
  TypeDescription: string;
  UOMCode: string;
  InvParentCateDescription: string;
  ClassGroupName: string;
  ProductType: null | string;
  MasterItemId: number;
  ItemSpecification: string | number;
  UomLookUpId: number;
  ItemAliasName: string;
  PurchaseTypeLookUpId: number;
  IsExpirable: boolean;
  ShelfLife: number;
  IsDiscountable: boolean;
  SaleTaxPurPercent: number;
  AtRetailPriceTax: boolean;
  AtPurcahsePriceTax: boolean;
  SaleTaxSalesPercent: number;
  IsTaxable: boolean;
  DepreciationExpenseAcId: number;
  accumulatedDepreciationAcId: number;
  CapitalWipAcId: number;
  EntryUserName: string;
  EntryUserId: number;
  EntryDate: Date;
  EntryUserProfileImage: null | string;
  MoidifyUserProfileImage: null | string;
  NoOfAttachments: number;
};